app/boat/requirements.txt
paho-mqtt==1.6.1
geopy==2.3.0